Refactor reindex modal
