+++ webapp/src/app/rootReducer.ts
@@ -7,7 +7,6 @@ import masterNodesReducer from '../containers/MasternodesPage/reducer';
 import syncStatusReducer from '../containers/SyncStatus/reducer';
 import cliReducer from '../containers/ConsolePage/reducer';
 import errorModalReducer from '../containers/ErrorModal/reducer';
-import reindexModelReducer from '../containers/ReIndexModel/reducer';
 import { i18nReducer } from 'react-redux-i18n';
 
 export default combineReducers({
@@ -20,5 +19,4 @@ export default combineReducers({
   i18n: i18nReducer,
   cli: cliReducer,
   errorModal: errorModalReducer,
-  reindexModel: reindexModelReducer,
 });
