--- webapp/src/containers/ReIndexModel/reindexModal.module.scss
@@ -1,6 +1,0 @@
-// App and Bootstrap variables required
-@import '~bootstrap/scss/functions';
-@import '~bootstrap/scss/variables';
-@import '~bootstrap/scss/mixins';
-@import '../../scss/variables';
-@import '../../scss/bootstrap/variables';
